Extract discord message content creation functionality into utility function
Fixes #6

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ppicazo/rage-against-the-meshine-beta/issues/6?shareId=XXXX-XXXX-XXXX-XXXX).
