+++ index.ts
@@ -24,6 +24,8 @@ import meshRedis from "./src/MeshRedis";
 import logger from "./src/Logger";
 import Commands from "./src/Commands";
 import { nodeHex2id, nodeId2hex } from "./src/NodeUtils";
+import { createDiscordMessage } from "./src/DiscordMessageUtils";
+
 
 // generate a pseduo uuid kinda thing to use as an instance id
 const INSTANCE_ID = (() => {
@@ -610,7 +612,7 @@ client.once("ready", () => {
     const balloonNode = await meshRedis.isBalloonNode(nodeId);
     // const trackerNode = await meshRedis.isTrackerNode(nodeId);
 
-    const content = await createDiscordMessage(packetGroup, text, balloonNode);
+    const content = await createDiscordMessage(packetGroup, text, balloonNode, client, guild);
 
     const getDiscordChannel = async (balloonNode, channelId) => {
       if (balloonNode) {
@@ -668,426 +670,6 @@ client.once("ready", () => {
       }
       // store message id in cache
       discordMessageIdCache.set(packet.id.toString(), discordMessage.id);
-    }
-  };
-
-  const createDiscordMessage = async (packetGroup, text, balloonNode) => {
-    try {
-      const packet = packetGroup.serviceEnvelopes[0].packet;
-      const from = nodeId2hex(packet.from);
-      const nodeIdHex = nodeId2hex(from);
-      const portNum = packet?.decoded?.portnum;
-      let msgText = text;
-
-      let nodeInfos = await meshRedis.getNodeInfos(
-        packetGroup.serviceEnvelopes
-          .map((se) => se.gatewayId.replace("!", ""))
-          .concat(from),
-        false,
-      );
-
-      let avatarUrl = "https://cdn.discordapp.com/embed/avatars/0.png";
-      // const discordPfpUrl = await meshRedis.getPfpUrl(nodeIdHex);
-      // if (discordPfpUrl) {
-      //   avatarUrl = discordPfpUrl;
-      // }
-
-      const maxHopStart = packetGroup.serviceEnvelopes.reduce((acc, se) => {
-        const hopStart = se.packet.hopStart;
-        return hopStart > acc ? hopStart : acc;
-      }, 0);
-
-      // logger.info("maxHopStart", maxHopStart);
-
-      const discordUserId = await meshRedis.getDiscordUserId(nodeIdHex);
-      logger.info(`nodeIdHex: ${nodeIdHex}, discordUserId: ${discordUserId}`);
-      let ownerField;
-      if (discordUserId) {
-        let guildUser: GuildMember | DiscordUser | undefined;
-        const user: DiscordUser = await client.users.fetch(discordUserId);
-        try {
-          guildUser = await guild.members.fetch(discordUserId);
-        } catch (e) {
-          logger.error(e);
-        }
-        if (!guildUser) {
-          logger.error(
-            `User ${discordUserId} not found in guild, using global user.`,
-          );
-          guildUser = user;
-        }
-        const userAvatarUrl = guildUser.displayAvatarURL();
-        if (userAvatarUrl && userAvatarUrl.length > 0) {
-          avatarUrl = userAvatarUrl;
-        }
-        // if (user.username === user.displayName) {
-        //   ownerField = {
-        //     name: "Owner",
-        //     value: `<@${user.id}>`,
-        //     inline: false,
-        //   };
-        // } else {
-        //   ownerField = {
-        //     name: "Owner*",
-        //     value: `<@!${user.id}>`,
-        //     inline: false,
-        //   };
-        // }
-        ownerField = {
-          name: "Owner",
-          value: userMention(user.id),
-          inline: false,
-        };
-      }
-
-      const gatewayCount = packetGroup.serviceEnvelopes.filter(
-        (value, index, self) =>
-          self.findIndex((t) => t.gatewayId === value.gatewayId) === index,
-      ).length;
-
-      logger.info(`gatewayCount: ${gatewayCount}`);
-
-      const infoFields: any = [];
-
-      let mapUrl = "";
-
-      if (portNum === 3) {
-        // https://api.bortle.org/api/maps/static?lat=32&lon=-122.0&width=800&height=600&zoom=6
-
-        const position = Position.decode(
-          packetGroup.serviceEnvelopes[0].packet.decoded.payload,
-        ) as DecodedPosition;
-        // logger.info(position);
-        // Position {
-        //   latitudeI: 379260350,
-        //   longitudeI: -1225297610,
-        //   altitude: 25,
-        //   time: 1743031998,
-        //   locationSource: 1,
-        //   groundSpeed: 0,
-        //   groundTrack: 0,
-        //   precisionBits: 32
-        // }
-        // envelope.packet.decoded.payload = `Latitude: ${position.latitudeI / 10000000} Longitude: ${position.longitudeI / 10000000} ${position.altitude ? ` Altitude: ${position.altitude}m` : ""}`;
-        // logger.info(from, envelope.packet.decoded.payload);
-
-        infoFields.push({
-          name: "Latitude",
-          value: `${position.latitudeI / 10000000}`,
-          inline: true,
-        });
-        infoFields.push({
-          name: "Longitude",
-          value: `${position.longitudeI / 10000000}`,
-          inline: true,
-        });
-        if (position.altitude) {
-          infoFields.push({
-            name: "Altitude",
-            value: `${position.altitude}m`,
-            inline: true,
-          });
-        }
-        // if (position.groundSpeed) {
-        //   infoFields.push({
-        //     name: "Altitude",
-        //     value: `${position.altitude}m`,
-        //     inline: true,
-        //   });
-        // }
-        // if (position.groundSpeed) {
-        //   infoFields.push({
-        //     name: "Altitude",
-        //     value: `${position.altitude}m`,
-        //     inline: true,
-        //   });
-        // }
-        // if (position.groundSpeed) {
-        //   infoFields.push({
-        //     name: "Altitude",
-        //     value: `${position.altitude}m`,
-        //     inline: true,
-        //   });
-        // }
-
-        logger.info(position);
-
-        try {
-          msgText = decodedPositionToString(position);
-        } catch (e) {
-          logger.error(e);
-        }
-        mapUrl = `https://api.smerty.org/api/v1/maps/static?lat=${position.latitudeI / 10000000}&lon=${position.longitudeI / 10000000}&width=400&height=400&zoom=12`;
-      }
-
-      logger.info(mapUrl);
-
-      if (ownerField) {
-        infoFields.push({
-          name: ownerField.name,
-          value: ownerField.value,
-          inline: ownerField.inline,
-        });
-      }
-
-      infoFields.push({
-        name: "Packet",
-        value: `[${packetGroup.id.toString(16)}](https://meshview.bayme.sh/packet/${packetGroup.id})`,
-        inline: true,
-      });
-
-      if (balloonNode) {
-        infoFields.push({
-          name: "Channel",
-          value: `${packetGroup.serviceEnvelopes[0].channelId}`,
-          inline: true,
-        });
-      }
-
-      infoFields.push({
-        name: "Hop Limit",
-        value: `${maxHopStart}`,
-        inline: true,
-      });
-      infoFields.push({
-        name: "Gateway Count",
-        value: `${gatewayCount}`,
-        inline: true,
-      });
-
-      // const gatewayFields = packetGroup.serviceEnvelopes
-      //   .filter(
-      //     (value, index, self) =>
-      //       self.findIndex((t) => t.gatewayId === value.gatewayId) === index,
-      //   )
-      //   .map((envelope) => {
-      //     const gatewayDelay =
-      //       envelope.mqttTime.getTime() - packetGroup.time.getTime();
-
-      //     let gatewayDisplaName = envelope.gatewayId.replace("!", "");
-      //     if (nodeInfos[envelope.gatewayId.replace("!", "")]) {
-      //       gatewayDisplaName =
-      //         nodeInfos[envelope.gatewayId.replace("!", "")].shortName;
-      //     }
-
-      //     let hopText = `${envelope.packet.hopStart - envelope.packet.hopLimit}/${envelope.packet.hopStart} hops`;
-
-      //     if (
-      //       envelope.packet.hopStart === 0 &&
-      //       envelope.packet.hopLimit === 0
-      //     ) {
-      //       hopText = `${envelope.packet.rxSnr} / ${envelope.packet.rxRssi} dBm`;
-      //     } else if (
-      //       envelope.packet.hopStart - envelope.packet.hopLimit ===
-      //       0
-      //     ) {
-      //       hopText = `${envelope.packet.rxSnr} / ${envelope.packet.rxRssi} dBm ${envelope.packet.hopStart - envelope.packet.hopLimit}/${envelope.packet.hopStart} hops`;
-      //     }
-
-      //     if (envelope.gatewayId.replace("!", "") === nodeIdHex) {
-      //       hopText = `Self Gated ${envelope.packet.hopStart} hopper`;
-      //     }
-
-      //     if (maxHopStart !== envelope.packet.hopStart) {
-      //       hopText = `:older_man: ${envelope.packet.hopStart - envelope.packet.hopLimit}/${envelope.packet.hopStart} hops`;
-      //     }
-
-      //     if (envelope.mqttServer === "public") {
-      //       hopText = `:poop: ${envelope.packet.hopStart - envelope.packet.hopLimit}/${envelope.packet.hopStart} hops`;
-      //     }
-
-      //     return {
-      //       name: `Gateway`,
-      //       value: `[${gatewayDisplaName} (${hopText})](https://meshview.bayme.sh/packet_list/${nodeHex2id(envelope.gatewayId.replace("!", ""))})${gatewayDelay > 0 ? " (" + gatewayDelay + "ms)" : ""}`,
-      //       inline: true,
-      //     };
-      //   });
-
-      // First, group gateways by hop count.
-      const gatewayGroups = {};
-
-      packetGroup.serviceEnvelopes
-        .filter(
-          (value, index, self) =>
-            self.findIndex((t) => t.gatewayId === value.gatewayId) === index,
-        )
-        .forEach((envelope) => {
-          const gatewayDelay =
-            envelope.mqttTime.getTime() - packetGroup.time.getTime();
-          let gatewayDisplayName = envelope.gatewayId.replace("!", "");
-          if (nodeInfos[gatewayDisplayName]) {
-            gatewayDisplayName = nodeInfos[gatewayDisplayName].shortName;
-          }
-
-          // Calculate the hop text based on several conditions.
-          let hopText;
-          if (
-            typeof envelope.packet.hopStart === "number" &&
-            typeof envelope.packet.hopLimit === "number"
-          ) {
-            hopText = ``;
-            if (
-              envelope.packet.hopStart === 0 &&
-              envelope.packet.hopLimit === 0
-            ) {
-              hopText = `(${envelope.packet.rxSnr} / ${envelope.packet.rxRssi} dBm)`;
-            } else if (
-              envelope.packet.hopStart - envelope.packet.hopLimit ===
-              0
-            ) {
-              hopText = `(${envelope.packet.rxSnr} / ${envelope.packet.rxRssi} dBm)`;
-            }
-            if (envelope.gatewayId.replace("!", "") === nodeIdHex) {
-              hopText = `(Self Gated)`;
-            }
-            if (maxHopStart !== envelope.packet.hopStart) {
-              // hopText = `(:older_man: ${envelope.packet.hopStart - envelope.packet.hopLimit}/${envelope.packet.hopStart} hops)`;
-            }
-            if (envelope.mqttServer === "public") {
-              // hopText = `(:poop: ${envelope.packet.hopStart - envelope.packet.hopLimit}/${envelope.packet.hopStart} hops)`;
-            }
-          } else {
-            hopText = "Unknown";
-          }
-
-          // Determine the grouping key.
-          // Here we use (hopStart - hopLimit) if available, otherwise "Unknown Hops".
-          let hopGroup;
-          if (
-            typeof envelope.packet.hopStart === "number" &&
-            typeof envelope.packet.hopLimit === "number" &&
-            maxHopStart === envelope.packet.hopStart
-          ) {
-            hopGroup = envelope.packet.hopStart - envelope.packet.hopLimit;
-          } else {
-            hopGroup = "Unknown Hops";
-          }
-
-          // Create the gateway field text.
-          // const gatewayFieldText =
-          //   `[${gatewayDisplayName} ${hopText}` +
-          //   (gatewayDelay > 0 ? `  - ${gatewayDelay}ms` : "") +
-          //   `](https://meshview.bayme.sh/packet_list/${nodeHex2id(envelope.gatewayId.replace("!", ""))})`;
-          const gatewayFieldText =
-            `[${gatewayDisplayName} ${hopText}` +
-            `](https://meshview.bayme.sh/packet_list/${nodeHex2id(envelope.gatewayId.replace("!", ""))})`;
-
-          // Group the text.
-          if (!gatewayGroups[hopGroup]) {
-            gatewayGroups[hopGroup] = [];
-          }
-          gatewayGroups[hopGroup].push(gatewayFieldText);
-        });
-
-      // Now, create one field per hop group. Unknown hops are placed last.
-      // const gatewayFields2 = Object.keys(gatewayGroups)
-      //   .sort((a, b) => {
-      //     if (a === "Unknown Hops") return 1;
-      //     if (b === "Unknown Hops") return -1;
-      //     return a - b;
-      //   })
-      //   .map((hop) => ({
-      //     name:
-      //       hop === "Unknown Hops"
-      //         ? "Unknown Hops"
-      //         : hop === "0"
-      //           ? "Direct"
-      //           : `${hop} hops`,
-      //     value: gatewayGroups[hop].join("\n"),
-      //     inline: false,
-      //   }));
-
-      // Now, create one field per hop group, splitting if a field's value exceeds 1024 characters.
-      const gatewayFields2: any = [];
-      Object.keys(gatewayGroups)
-        .sort((a, b) => {
-          if (a === "Unknown Hops") return 1;
-          if (b === "Unknown Hops") return -1;
-          return a - b;
-        })
-        .forEach((hop) => {
-          // Set a base name for the field
-          const baseName =
-            hop === "Unknown Hops"
-              ? "Unknown Hops"
-              : hop === "0"
-                ? "Direct"
-                : `${hop} hops`;
-
-          // Prepare to build the field text by iterating over each line (gatewayFieldText) in the group.
-          const lines = gatewayGroups[hop];
-          let currentChunk = "";
-          let fieldIndex = 0;
-
-          lines.forEach((line) => {
-            // +1 accounts for the newline if currentChunk isn't empty.
-            if (
-              currentChunk.length +
-                line.length +
-                (currentChunk.length > 0 ? 1 : 0) >
-              1024
-            ) {
-              // Push the current chunk as a field.
-              gatewayFields2.push({
-                name: fieldIndex === 0 ? baseName : `${baseName} continued`,
-                value: currentChunk,
-                inline: false,
-              });
-              fieldIndex++;
-              // Start a new chunk with the current line.
-              currentChunk = line;
-            } else {
-              // Append the line, adding a newline if needed.
-              currentChunk =
-                currentChunk.length > 0
-                  ? currentChunk + (hop === "0" ? "\n" : " | ") + line
-                  : line;
-            }
-          });
-
-          // Push any remaining text in the current chunk.
-          if (currentChunk.length > 0) {
-            gatewayFields2.push({
-              name: fieldIndex === 0 ? baseName : `${baseName} (continued)`,
-              value: currentChunk,
-              inline: false,
-            });
-          }
-        });
-
-      const content = {
-        username: "Mesh Bot",
-        avatar_url:
-          "https://cdn.discordapp.com/app-icons/1240017058046152845/295e77bec5f9a44f7311cf8723e9c332.png",
-        embeds: [
-          {
-            url: `https://meshview.bayme.sh/packet_list/${packet.from}`,
-            color: 6810260,
-            timestamp: new Date(packet.rxTime * 1000).toISOString(),
-
-            author: {
-              name: `${nodeInfos[nodeIdHex] ? nodeInfos[nodeIdHex].longName : "Unknown"}`,
-              url: `https://meshview.bayme.sh/packet_list/${packet.from}`,
-              icon_url: avatarUrl,
-            },
-            title: `${nodeInfos[nodeIdHex] ? nodeInfos[nodeIdHex].shortName : "UNK"}`,
-            description: msgText,
-            fields: [...infoFields, ...gatewayFields2].slice(0, 25),
-          },
-        ],
-      };
-
-      if (mapUrl) {
-        content.embeds[0].image = {
-          url: mapUrl,
-        };
-      }
-      // logger.info(content);
-
-      return content;
-    } catch (err) {
-      logger.error("Error: " + String(err));
-      // Sentry.captureException(err);
     }
   };
